Add Git provider login handlers
